refactored the localization logic into it's own class
